Move source from lib to src
