+++ package.json
@@ -13,10 +13,10 @@
   "contributors": [
     "Adriean Khisbe <adriean.khisbe@live.fr> (https://github.com/AdrieanKhisbe/)"
   ],
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "mocha",
-    "lint": "eslint lib test",
+    "lint": "eslint src test",
     "integration-tests": "(cd examples && npm run test)"
   },
   "repository": {
